agents/windows_ad/service/requirements.txt
httpx
ldap3
pyyaml
python-dotenv
pywin32; platform_system == "Windows"
